pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "setuptools-scm>=8.0"
  ]
build-backend = "setuptools.build_meta"


[project]
name = "cyberclip"
version = "0.0.3"
authors = [
  { name="BongoKnight" },
]
description = "A parser for handling recognisable data such as IP, domain, MD5 and process them with custom actions."
readme = "README.md"
requires-python = ">=3.10,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "PyYAML",
  "XlsxWriter",
  "beautifulsoup4",
  "dnspython",
  "jsonpath",
  "jq",
  "pandas",
  "pycti",
  "pywin32; platform_system=='Windows'",
  "pyperclip",
  "requests",
  "shodan",
  "dotenv",
  "textual",
  "textual[syntax]",
  "textual-fspicker",
  "urllib3",
  "tldextract"
]

[project.optional-dependencies]
# Radial GUI menu (Qt-based) - Windows-focused
gui = [
  "PyQt5>=5.15.0",
  "keyboard>=0.13.0",
  "win11toast>=0.34; platform_system=='Windows'"
]

# Web interface (Flask-based)
web = [
  "flask>=2.0.0",
  "textual_serve"
]

# All optional features
all = [
  "PyQt5>=5.15.0",
  "keyboard>=0.13.0",
  "win11toast>=0.34; platform_system=='Windows'",
  "flask>=2.0.0",
  "textual_serve"
]

[tool.setuptools.package-data]
"cyberclip.data" = ["*.bak"]

[project.scripts]
# Main unified entry point with interface selector
cyberclip = "cyberclip.__main__:main"

# Direct interface launchers
cyberclip-tui = "cyberclip.app:main"
cyberclip-gui = "cyberclip.radialUI:main"
cyberclip-web = "cyberclip.flask_app:main"

[project.urls]
Homepage = "https://github.com/BongoKnight/cyberclip"
Issues = "https://github.com/BongoKnight/cyberclip/issues"
